requirements.txt
amqplib==1.0.2
click==6.6
docopt==0.6.2
Flask==0.11.1
gunicorn==19.6.0
httplib2==0.9.2
itsdangerous==0.24
Jinja2==2.8
MarkupSafe==0.23
MySQL-python==1.2.5
protobuf==2.4.0a0
PyJWT==1.4.1
python-ldap==2.4.27
python-memcached==1.58
pytz==2016.6.1
six==1.10.0
Twisted==16.3.0
txAMQP==0.6.2
tzlocal==1.2.2
urllib3==1.16
Werkzeug==0.11.10
zenoss.protocols==0.2
zope.component==4.2.2
zope.dottedname==4.1.0
zope.event==4.2.0
zope.interface==4.2.0
